feat: update EditingBlock to follow structure as ViewerBlock.
