+++ app/Component/Editing/MarkEditingBlock/styles.module.css
@@ -1,0 +1,27 @@
+.boxContent {
+  display: flex;
+  justify-content: space-between;
+  flex-direction: column;
+  background-color: #fdfdfd;
+}
+
+.boxContentDraft {
+  max-height: 44vh;
+  min-height: 35vh;
+  box-sizing: border-box;
+  padding: 5vh 10% 0 6.5%;
+  overflow-y: auto;
+  cursor: text;
+}
+
+.boxDraftBottom{
+  margin: 2.7vh 0 1.4vh 0;
+}
+
+.fontContentDraft {
+  font-size: 1.36rem;
+  letter-spacing: 0.1rem;
+  line-height: 1.7rem;
+  font-weight: 400;
+  color: #000000;
+}
