+++ app/Within/stylesCosmic.module.css
@@ -7,8 +7,8 @@
 }
 
 .fontExplore {
-  font-size: 1.24rem;
-  letter-spacing: 0.1rem;
+  font-size: 1.27rem;
+  letter-spacing: 0.12rem;
   white-space: nowrap;
   color: #353535;
 }
