+++ public/css/stylesSelfFront.css
@@ -60,6 +60,33 @@
   line-height: 1.7rem;
   color: #ABABAB;
 }
+._C_C_app_Component_Editing_MarkEditingBlock_styles_module__boxContent {
+  display: flex;
+  justify-content: space-between;
+  flex-direction: column;
+  background-color: #fdfdfd;
+}
+
+._C_C_app_Component_Editing_MarkEditingBlock_styles_module__boxContentDraft {
+  max-height: 44vh;
+  min-height: 35vh;
+  box-sizing: border-box;
+  padding: 5vh 10% 0 6.5%;
+  overflow-y: auto;
+  cursor: text;
+}
+
+._C_C_app_Component_Editing_MarkEditingBlock_styles_module__boxDraftBottom{
+  margin: 2.7vh 0 1.4vh 0;
+}
+
+._C_C_app_Component_Editing_MarkEditingBlock_styles_module__fontContentDraft {
+  font-size: 1.36rem;
+  letter-spacing: 0.1rem;
+  line-height: 1.7rem;
+  font-weight: 400;
+  color: #000000;
+}
 ._C_C_app_Component_Unit_Mark_AuthorBlock_styles_module__boxPanelJump {
   box-sizing: border-box;
   padding: 0 10% 0 8%;
